Move analyze.py from Plotter to plotter.py in drivers
